A couple graphic fixes + Begin Reimplementing Style Select
Due to ScreenSelectMaster being a bit limited, I've begun doing a re-write which'll allow us to slim down the amount of extra code needed for what we need to do.
It currently only shows a counter on screen if debug is enabled but I'll start adding stuff as time goes on.